# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "图书馆管理系统API文档",
        "title": "Library Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/books": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "根据条件搜索图书",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "获取图书列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "true: 只显示可借阅的图书",
                        "name": "available",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "minLength": 1,
                        "type": "string",
                        "example": "Fiction",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "minLength": 1,
                        "type": "string",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "number",
                        "example": 20,
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "number",
                        "example": 10,
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            0,
                            1
                        ],
                        "type": "integer",
                        "example": 0,
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员创建新图书，包含ISBN、书名、作者等基本信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "创建新图书",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "图书信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功创建图书",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Book"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取指定图书的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "获取图书详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "图书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员更新图书信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "更新图书信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "图书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "图书信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/books/{id}/status": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员更新图书上下架状态 2-下架 1-上架",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "更新图书状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "图书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.StatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/books/{id}/stock": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员更新图书库存数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "图书管理"
                ],
                "summary": "更新图书库存",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户的访问令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "图书ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "库存信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateBookStockRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Book": {
            "description": "图书信息",
            "type": "object",
            "properties": {
                "author": {
                    "description": "作者",
                    "type": "string"
                },
                "available": {
                    "description": "可借数量",
                    "type": "integer"
                },
                "category": {
                    "description": "分类",
                    "type": "string"
                },
                "cover": {
                    "description": "封面图片URL",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "删除时间",
                    "type": "string",
                    "format": "date-time",
                    "example": "2024-01-01T00:00:00+08:00"
                },
                "id": {
                    "description": "bookID",
                    "type": "integer"
                },
                "isbn": {
                    "description": "ISBN编号",
                    "type": "string"
                },
                "location": {
                    "description": "馆藏位置",
                    "type": "string"
                },
                "price": {
                    "description": "价格",
                    "type": "number"
                },
                "publisher": {
                    "description": "出版社",
                    "type": "string"
                },
                "status": {
                    "description": "状态 2-下架 1-上架",
                    "type": "integer"
                },
                "summary": {
                    "description": "简介",
                    "type": "string"
                },
                "title": {
                    "description": "书名",
                    "type": "string"
                },
                "total": {
                    "description": "总数量",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "request.CreateBookRequest": {
            "description": "创建新图书的请求参数",
            "type": "object",
            "required": [
                "author",
                "category",
                "isbn",
                "location",
                "price",
                "publisher",
                "title",
                "total"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1,
                    "example": "J.D. Salinger"
                },
                "category": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "Fiction"
                },
                "cover": {
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "isbn": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10,
                    "example": "9787111111111"
                },
                "location": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "Shelf A1"
                },
                "price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 10
                },
                "publisher": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1,
                    "example": "Little, Brown and Company"
                },
                "summary": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "This is a great book about life and love."
                },
                "title": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "example": "The Catcher in the Rye"
                },
                "total": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 100
                }
            }
        },
...
        "request.StatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "2:禁用 1:启用",
                    "type": "integer"
                }
            }
        },
        "request.UpdateBookRequest": {
            "description": "更新图书信息的请求参数",
            "type": "object",
            "properties": {
                "author": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1,
                    "example": "J.D. Salinger"
                },
                "category": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "Fiction"
                },
                "cover": {
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "location": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1,
                    "example": "Shelf A1"
                },
                "price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 10
                },
                "publisher": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 1,
                    "example": "Little, Brown and Company"
                },
                "summary": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "This is a great book about life and love."
                },
                "title": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 1,
                    "example": "The Catcher in the Rye"
                },
                "total": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 100
                }
            }
        },
        "request.UpdateBookStockRequest": {
            "description": "更新图书库存的请求参数",
            "type": "object",
            "required": [
                "change"
            ],
            "properties": {
                "change": {
                    "description": "可以为负数，表示减少库存",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "response.Response": {
            "description": "API通用响应结构",
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}